QuantAgent/requirements.txt
anthropic>=0.39.0
ib_insync>=0.9.86
pyvinecopulib>=0.7.0
numpy>=1.26.0
scipy>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0
